backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.0.3
Jinja2==3.1.4
python-dotenv==1.0.1
psycopg2==2.9.10
WeasyPrint==66.0
gunicorn==22.0.0
openai==1.68.2
requests==2.32.3
